requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.1
configparser==6.0.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
packaging==23.2
paho-mqtt==2.0.0
parso==0.8.3
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyModbusTCP==0.2.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
uvicorn==0.27.1
wcwidth==0.2.13
zipp==3.17.0
